+++ Tests/UINotificationsTests/UINotificationTestCase.swift
@@ -5,6 +5,7 @@
 //  Created by Antoine van der Lee on 14/07/2017.
 //  Copyright © 2017 WeTransfer. All rights reserved.
 //
+//  danger:disable final_class
 
 import XCTest
 @testable import UINotifications
@@ -17,7 +18,7 @@ class UINotificationTestCase: XCTestCase {
         
         var dismissTrigger: UINotificationDismissTrigger
         var presentationContext: UINotificationPresentationContext
-        var isDismissing: Bool = false
+        var state: UINotificationPresenterState = .idle
         
         private(set) var presented: Bool = false
         private(set) var dismissed: Bool = false
@@ -30,11 +31,13 @@ class UINotificationTestCase: XCTestCase {
         func present() {
             dismissTrigger.target = self
             (dismissTrigger as? UINotificationSchedulableDismissTrigger)?.schedule()
+            state = .presented
             presented = true
         }
         
         func dismiss() {
             presentationContext.completePresentation()
+            state = .idle
             dismissed = true
         }
     }
@@ -44,6 +47,7 @@ class UINotificationTestCase: XCTestCase {
         var dismissTrigger: UINotificationDismissTrigger
         var presentationContext: UINotificationPresentationContext
         var isDismissing: Bool = false
+        var state: UINotificationPresenterState = .idle
         
         private(set) var presented: Bool = false
         private(set) var dismissed: Bool = false