+++ Changelog.md
@@ -8,6 +8,8 @@
 - Chevron image is now set inside the style, which is more logic.
 - The chevron image is now hidden when there's no action attached.
 - It's possible to override the default notification view.
+- Cancellation is now only possible when the notification request is not yet running.
+- Better state representation for `UINotificationPresenter`.
 
 ### 1.0 (2017-07-14)